nav_bar: make into icon and label own exported components
